requirements-test.txt
pytest==6.2.4
pytest-aiohttp==0.3.0
